requirements.txt
Django==1.6
Pillow==2.4.0
PyPDF2==1.23
South==0.8.4
argparse==1.2.1
diff-match-patch==20121119
dj-database-url==0.3.0
dj-static==0.0.6
django-extensions==1.4.0
django-grappelli==2.4.5
django-import-export==0.2.3
django-toolbelt==0.0.1
djangorestframework==2.4.2
gunicorn==19.1.1
html5lib==1.0b3
httplib2==0.9
psycopg2==2.5.4
qrcode==5.0.1
reportlab==3.1.8
six==1.7.3
static3==0.5.1
tablib==0.10.0
wsgiref==0.1.2
xhtml2pdf==0.0.6
